ClusterFitter: refactor so output code is seperate from fitter.
FossilOrigin-Name: 1dd3bce8282574d2e562e3e8edc0d1969fdbaab2e0814b62aec988fb88f2f4b7

+++ statfile/test/test_BinFormat.cpp
@@ -118,10 +118,10 @@ namespace {
 				str << (i+j) ;
 				switch( c ) {
 					case 0:
-						sink << int32_t( i+j ) ;
+						sink << int64_t( i+j ) ;
 						break ;
 					case 1:
-						sink << uint32_t( i+j ) ;
+						sink << int64_t( i+j ) ;
 						break ;
 					case 2:
 						sink << double( std::exp( i+j ) ) ;
